Moved ParameterGroupExtension to generic extensions
